Checkpoint: Code refactor
