+++ .pre-commit-config.yaml
@@ -1,14 +1,14 @@
 repos:
-- repo: local
-  hooks:
-    - id: nox
-      name: nox
-      files: ^.*\.py$
-      language: python
-      entry: nox
-      types: [python]
-      # stages: [pre-commit, push]
-      pass_filenames: false
+# - repo: local
+#   hooks:
+#     - id: nox
+#       name: nox
+#       files: ^.*\.py$
+#       language: python
+#       entry: nox
+#       types: [python]
+#       # stages: [pre-commit, push]
+#       pass_filenames: false
 
 # We don't need to do this seperately, because nox will run the tests
 - repo: local
